Add VS Code debug configuration (#48)
* Add vscode debug config for flask app

Adds a docker-compose.debug.yml file based on the existing docker-compose.yml file used in dev, which overrides the default command in order to install debugpy and wait for a debugger to attach via port 6789 prior to running the flask dev server process.

* Add vscode debug config for save_routes.py

Adds configuration for debugging the `save_routes.py` script using the flask-dev Docker image. This configuration can be used as prior art for adding other script debug configuration in the future, or adjusted to accept a user-specified command to run rather than creating vscode tasks for each script. Note: a prompt was added to allow M1/M2 Mac users to select a working platform; this will ideally be removed in the future as support for ARM64 wheels is added to installed packages, but is necessary for now to successfully build the Docker image.

* Add docs for new debug configuration

* Update compose debug config to be used as extension to base compose config

Updates the `docker-compose.debug.yml` config to define only settings that should override or extend the base `docker-compose.yml` config.